requirements.txt
rembg==2.0.66
pillow==10.4.0
